Some review changes, and changed snapshot version.
