+++ version.sbt
@@ -1,1 +1,1 @@
-version in ThisBuild := "1.2.0-SNAPSHOT"
+version in ThisBuild := "1.1.0-nosr-SNAPSHOT"
